refactor : UserDTO refactoring
