+++ src/main/java/com/mega/biz/join/model/UserDTODeserializer.java
@@ -1,19 +1,20 @@
-package com.mega.biz.join.model.dto;
+package com.mega.biz.join.model;
 
 import com.google.gson.*;
+import com.mega.biz.join.model.dto.UserDTO;
 
 import java.lang.reflect.Type;
 
-public class UserDTOvDeserializer implements JsonDeserializer<UserDTOv> {
+public class UserDTODeserializer implements JsonDeserializer<UserDTO> {
     @Override
-    public UserDTOv deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
+    public UserDTO deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
         JsonObject jsonObject = json.getAsJsonObject();
         String email = jsonObject.get("email").getAsString();
         String password = jsonObject.get("password").getAsString();
         String name = jsonObject.get("name").getAsString();
         String phone = jsonObject.get("phone").getAsString();
 
-        UserDTOv userDTO = new UserDTOv();
+        UserDTO userDTO = new UserDTO();
         userDTO.setEmail(email);
         userDTO.setPassword(password);
         userDTO.setName(name);
