+++ src/main/java/com/mega/biz/join/model/dto/UserDTO.java
@@ -1,18 +1,51 @@
 package com.mega.biz.join.model.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import com.mega.biz.join.model.domain.Email;
+import com.mega.biz.join.model.domain.Password;
+import com.mega.biz.join.model.domain.Phone;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
 
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
+@Getter
+@Setter
 public class UserDTO {
 
-    private String email;
-    private String password;
+    @Getter(AccessLevel.NONE)
+    @Setter(AccessLevel.NONE)
+    private Email email;
+    @Getter(AccessLevel.NONE)
+    @Setter(AccessLevel.NONE)
+    private Password password;
     private String name;
-    private String phone;
+    @Getter(AccessLevel.NONE)
+    @Setter(AccessLevel.NONE)
+    private Phone phone;
+    private String encrypedPassword;
     private Long userStatus;
     private String salt;
+
+    public void setEmail(String email) {
+        this.email = new Email(email);
+    }
+
+    public void setPassword(String password) {
+        this.password = new Password(password);
+    }
+
+    public void setPhone(String phone) {
+        this.phone = new Phone(phone);
+    }
+
+    public String getEmail() {
+        return email.getEmail();
+    }
+
+    public String getPassword() {
+        return password.getPassword();
+    }
+
+    public String getPhone() {
+        return phone.getPhone();
+    }
 }
