+++ src/main/java/com/mega/biz/join/model/dto/UserDTOOriginal.java
@@ -1,0 +1,18 @@
+package com.mega.biz.join.model.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+public class UserDTOOriginal {
+
+    private String email;
+    private String password;
+    private String name;
+    private String phone;
+    private Long userStatus;
+    private String salt;
+}
