+++ src/main/java/com/mega/biz/join/service/JoinService.java
@@ -1,8 +1,8 @@
 package com.mega.biz.join.service;
 
 import com.mega.biz.join.model.JoinDAO;
+import com.mega.biz.join.model.dto.UserDTOOriginal;
 import com.mega.biz.join.model.dto.UserDTO;
-import com.mega.biz.join.model.dto.UserDTOv;
 import com.mega.common.encrypt.EncryptUtils;
 
 public class JoinService {
@@ -12,19 +12,19 @@ public class JoinService {
     private final EncryptUtils encrypt = new EncryptUtils();
 
 
+//    public boolean saveUser(UserDTOOriginal userDTOOriginal) {
+//        String salt = encrypt.getSalt();
+//        userDTOOriginal.setSalt(salt);
+//
+//        String encryptedPassword = encrypt.getEncrypt(userDTOOriginal.getPassword(), salt);
+//        userDTOOriginal.setPassword(encryptedPassword);
+//        userDTOOriginal.setUserStatus(1L);
+//
+//        int i = dao.insertUser(userDTOOriginal);
+//        return i == 1;
+//    }
+
     public boolean saveUser(UserDTO userDTO) {
-        String salt = encrypt.getSalt();
-        userDTO.setSalt(salt);
-
-        String encryptedPassword = encrypt.getEncrypt(userDTO.getPassword(), salt);
-        userDTO.setPassword(encryptedPassword);
-        userDTO.setUserStatus(1L);
-
-        int i = dao.insertUser(userDTO);
-        return i == 1;
-    }
-
-    public boolean saveUser2(UserDTOv userDTO) {
         String salt = encrypt.getSalt();
         userDTO.setSalt(salt);
 
@@ -32,11 +32,11 @@ public class JoinService {
         userDTO.setEncrypedPassword(encryptedPassword);
         userDTO.setUserStatus(1L);
 
-        int i = dao.insertUser2(userDTO);
+        int i = dao.insertUser(userDTO);
         return i == 1;
     }
 
-    public UserDTO findUser(UserDTO userDTO) {
-        return dao.findUser(userDTO);
+    public UserDTOOriginal findUser(UserDTOOriginal userDTOOriginal) {
+        return dao.findUser(userDTOOriginal);
     }
 }
